requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.3.0
backoff==2.2.1
bert-score==0.3.13
bidict==0.23.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chainlit==2.4.400
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
dataclasses-json==0.6.7
deepeval==2.7.6
Deprecated==1.2.18
distro==1.9.0
docutils==0.21.2
execnet==2.1.1
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.11.0
googleapis-common-protos==1.69.2
groq==0.26.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.63
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-ollama==0.3.1
langchain-openai==0.3.12
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.30
Lazify==0.4.0
literalai==0.1.201
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.6.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==2.2.4
ollama==0.4.7
openai==1.76.0
opencv-python==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.0
opentelemetry-instrumentation-anthropic==0.39.0
opentelemetry-instrumentation-bedrock==0.39.0
opentelemetry-instrumentation-chromadb==0.39.0
opentelemetry-instrumentation-cohere==0.39.0
opentelemetry-instrumentation-crewai==0.39.0
opentelemetry-instrumentation-google-generativeai==0.39.0
opentelemetry-instrumentation-groq==0.39.0
opentelemetry-instrumentation-haystack==0.39.0
opentelemetry-instrumentation-lancedb==0.39.0
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-llamaindex==0.39.0
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.0
opentelemetry-instrumentation-milvus==0.39.0
opentelemetry-instrumentation-mistralai==0.39.0
opentelemetry-instrumentation-ollama==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-pinecone==0.39.0
opentelemetry-instrumentation-qdrant==0.39.0
opentelemetry-instrumentation-replicate==0.39.0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.0
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.0
opentelemetry-instrumentation-transformers==0.39.0
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.0
opentelemetry-instrumentation-watsonx==0.39.0
opentelemetry-instrumentation-weaviate==0.39.0
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
pkginfo==1.10.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
posthog==3.24.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.4.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.13.3
readme_renderer==44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.25.1
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentry-sdk==2.26.1
setuptools==78.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.41.3
streamlit==1.45.1
sympy==1.14.0
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.39.0
transformers==4.51.3
twine==5.1.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.149
ultralytics-thop==2.0.14
uptrace==1.31.0
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==0.20.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
